Module Js_of_ocaml

module CSS : sig ... end

This module contains a few types and values to ease the use of CSS properties and such. If you think a feature is missing, consider sending a patch or an RFE to the mailing list.

module Dom : sig ... end

DOM binding

module Dom_events : sig ... end

Javascript events

module Dom_html : sig ... end

DOM HTML binding

module Dom_svg : sig ... end

DOM SVG binding

module EventSource : sig ... end

EventSource binding

module File : sig ... end

File API

module Firebug : sig ... end

Firebug API (debugging console).

module Form : sig ... end

module Geolocation : sig ... end

Geolocation API

module Intl : sig ... end

Internationalization API

module Js : sig ... end

Javascript binding

module Json : sig ... end

Unsafe IO. (See Deriving_Json for typesafe IO)

module Jstable : sig ... end

A minimal table implementation specialized for Js.js_string keys. This is faster than regular OCaml hashtables.

module MutationObserver : sig ... end

MutationObserver API

module ResizeObserver : sig ... end

ResizeObserver API

module Regexp : sig ... end

Types for regexps.

module Sys_js : sig ... end

Javascript specific Sys functions.

module Typed_array : sig ... end

Typed Array binding

module Url : sig ... end

This module provides functions for tampering with Url. It's main goal is to allow one to stay in the Ocaml realm without wandering into the Dom_html.window##.location object.

module WebGL : sig ... end

WebGL binding

module WebSockets : sig ... end

WebSocket binding

module Worker : sig ... end

Low-level bindgins to javascript Web Workers.

module XmlHttpRequest : sig ... end

XmlHttpRequest object.