Module Js_of_ocaml.IntersectionObserver

The Intersection Observer API provides a way to asynchronously observe changes in the intersection of a target element with an ancestor element or with a top-level document's viewport.

https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API

class type intersectionObserverEntry = object ... end

class type intersectionObserverOptions = object ... end

class type intersectionObserver = object ... end

val empty_intersection_observer_options : unit -> intersectionObserverOptions Js.t

val is_supported : unit -> bool

val intersectionObserver : ((intersectionObserverEntry Js.t Js.js_array Js.t -> intersectionObserver Js.t -> unit) Js.callback -> intersectionObserverOptions Js.t -> intersectionObserver Js.t) Js.constr